fix missing materials, and account for RCSG folder being moved when generating RP materials.
instead of using the shader name to generate folders, it now uses URP and HDRP respectively (based on the default shader present in the RP) when creating RP materials, to reduce folder structure depth. additionally, RCSG now uses an assembly definition, which is used to locate RCSG when locating assets.